fix(docs): add expand animation
